llm-service/requirements.txt
torch
transformers
fastapi
pydantic
uvicorn
requests
pillow
